Decreased overdraw.
Removed a lot of drawings that were unnecessary.
Updated the setlist images to be darker to begin with.
TODO: Revert the images and use code to darken, now that I know how.
Changed cached background images to Bitmap instead of BitmapDrawable to
make them easier to manipulate.

Attempted to override onDraw in subclassed ImageView for background to
reduce some overdraws, to no effect, so reverted with commented code
still present.